Move `Reset`trait from `mithril-common` to `mithril-resource-pool`
