+++ internal/mithril-resource-pool/src/resource_pool.rs
@@ -9,7 +9,10 @@ use std::{
 };
 use thiserror::Error;
 
-use mithril_common::{Reset, StdResult};
+use mithril_common::{
+    crypto_helper::{MKMap, MKMapKey, MKMapValue, MKTreeStorer},
+    StdResult,
+};
 
 /// [ResourcePool] related errors.
 #[derive(Error, Debug)]
@@ -184,6 +187,12 @@ impl<'a, T: Reset + Send + Sync> ResourcePoolItem<'a, T> {
     }
 }
 
+impl<K: MKMapKey, V: MKMapValue<K>, S: MKTreeStorer> Reset for MKMap<K, V, S> {
+    fn reset(&mut self) -> StdResult<()> {
+        self.compress()
+    }
+}
+
 impl<T: Reset + Send + Sync> Deref for ResourcePoolItem<'_, T> {
     type Target = T;
 
@@ -206,6 +215,19 @@ impl<T: Reset + Send + Sync> Drop for ResourcePoolItem<'_, T> {
         });
     }
 }
+
+/// Reset trait which is implemented by pooled resource items.
+/// As pool resource items are mutable, this will guarantee that the pool stays  consistent
+/// and that acquired resource items do not depend from previous computations.
+pub trait Reset {
+    /// Reset the resource
+    fn reset(&mut self) -> StdResult<()> {
+        Ok(())
+    }
+}
+
+#[cfg(test)]
+impl Reset for String {}
 
 #[cfg(test)]
 mod tests {
